collector/requirements.txt
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
attrs==25.4.0
av==15.1.0
cloudpickle==3.1.2
cmake==4.1.3
cmeel==0.59.0
cmeel-assimp==6.0.2
cmeel-boost==1.89.0
cmeel-console-bridge==1.0.2.3
cmeel-octomap==1.10.0
cmeel-qhull==8.0.2.1
cmeel-tinyxml2==10.0.0
cmeel-urdfdom==4.0.1
cmeel-zlib==1.3.1
coal==3.0.2
contourpy==1.3.3
cuda-bindings==12.9.4
cuda-pathfinder==1.4.2
cycler==0.12.1
datasets==4.7.0
deepdiff==8.6.1
diffusers==0.35.2
dill==0.4.0
docstring_parser==0.17.0
draccus==0.10.0
dynamixel-sdk==4.0.3
eigenpy==3.12.0
einops==0.8.2
evdev==1.9.3
Farama-Notifications==0.0.4
feetech-servo-sdk==1.0.0
filelock==3.25.1
fonttools==4.62.0
frozenlist==1.8.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.30.0
googleapis-common-protos==1.73.0
google-auth==2.49.0
google-cloud-aiplatform==1.141.0
google-cloud-bigquery==3.40.1
google-cloud-core==2.5.0
google-cloud-resource-manager==1.16.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-genai==1.66.0
google-resumable-media==2.8.0
grpc-google-iam-v1==0.14.3
grpcio==1.78.0
grpcio-status==1.78.0
gymnasium==1.2.3
hf-xet==1.3.2
huggingface_hub==1.6.0
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.1
Jinja2==3.1.6
jsonlines==4.0.0
kiwisolver==1.5.0
libcoal==3.0.2
libpinocchio==3.9.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mergedeep==1.3.4
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
mypy_extensions==1.1.0
networkx==3.6.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
openai>=1.0.0
opencv-contrib-python==4.13.0.92
orderly-set==5.5.0
pandas==3.0.1
pillow==12.1.1
pin==3.9.0
propcache==0.4.1
protobuf==6.33.5
proto-plus==1.27.1
psutil==7.2.2
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pynput==1.8.1
pyparsing==3.3.2
pyrealsense2==2.56.5.9235
pyserial==3.5
python-dateutil==2.9.0.post0
python-xlib==0.33
PyYAML==6.0.3
pyyaml-include==1.4.1
regex==2026.2.28
rerun-sdk==0.26.2
rsa==4.9.1
safetensors==0.7.0
scipy==1.17.1
sentry-sdk==2.54.0
setuptools==80.10.2
six==1.17.0
smmap==5.0.3
sympy==1.14.0
tenacity==9.1.4
termcolor==3.3.0
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0
torchcodec==0.10.0
torchvision==0.25.0
tqdm>=4.65.0
transformers==5.3.0
trimesh==4.11.3
triton==3.6.0
typer==0.20.0
typer-slim==0.20.0
typing-inspect==0.9.0
wandb==0.24.2
websockets==16.0
wheel==0.46.3
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
